TotalWeight should be a long instead of an int.

+++ ShopifySharp/Entities/ShopifyOrder.cs
@@ -270,7 +270,7 @@ namespace ShopifySharp
         /// The sum of all the weights of the line items in the order, in grams.
         /// </summary>
         [JsonProperty("total_weight")]
-        public int? TotalWeight { get; set; }
+        public long? TotalWeight { get; set; }
 
         /// <summary>
         /// The date and time when the order was last modified.
